Migrate service to split utils
